refactor: changed staking default value
Previously the UI suggested the user to stake the max amount of PNK available.

This led to a problem where users would be automatically unstaked from all
courts if they lost a case, because the slashing would make their PNK
balance to be lower than the amount staked.

This change introduce a suggested "buffer" of 2000 PNK to allow coping with losing
a case and not being unstaked.

It simply warns users. If they choose to ignore the warning, they can
still stake the full available balance.