feat: added link to get side-chain PNK
Added a link to Home and Courts to help users to get additional
side-chain PNK when they are on a side-chain.

The process requires them to switch back to an Ethereum network
to make the swap. The user will be redirected to the Token Bridge
page with the proper required chain ID parameter set.