feat: store committed votes locally
The Court UI currently does not store the committed vote anywhere when
hidden votes are enabled.

The interface can become a bit confusing, because vote and reveal would
show the same options and the jurors are required to select the same
options they selected in the first place. If they choose wrong, the
transaction will fail (MetaMask will show a warning at least).

Since hidden votes are enabled on xDAI General Court, I implemented some
changes in the voting widget that will allow to store the committed vote
locally and present it to the user. To reveal, instead of selecting all
options over again, it simply submits the stored vote.

Notice that this will only work if the jurors reveal their votes using
the same device they used to commit to that specific case. In case the
UI cannot find the committed vote stored locally, it will fallback to
the current behavior, requiring the jurors to select the option(s) they
committed to.