feat: added algorithm to derive the committed vote
In case jurors tries to reveal their votes from a device different from
the one they committed to it, we can still try to derive the vote for
cases whose answer type is either `single-select` or `multiple-select`.

This can be done through brute force, generating every possible vote
and trying to match it against the value of the on-chain commit.