fix: show justifications for the ongoing round
there seemed to be a deliberate choice
to hide the justifications for a vote
in the current ongoing round. im
not sure why.
it lies to the users (it will claim that there
are no justifications, which isn't true) and it's not obvious
why the current justifications should be hidden,
more so if the data is publicly available.
i guess rationale was to not sway jurors.

technical users can get that info from the endpoint
anyway. or they can read the network log.